requirements.txt
rich
dnspython
requests
pyfiglet